requirements.txt
python-dotenv
openai
line-bot-sdk
awscli
boto3